+++ src/main/java/com/api/ttoklip/domain/aop/notification/annotation/SendNotification.java
@@ -1,4 +1,4 @@
-package com.api.ttoklip.domain.common.filtering.aop.annotation;
+package com.api.ttoklip.domain.aop.notification.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -7,5 +7,5 @@ import java.lang.annotation.Target;
 
 @Target(ElementType.METHOD)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface CheckBadWordUpdate {
+public @interface SendNotification {
 }
