+++ src/main/java/com/api/ttoklip/domain/honeytip/post/service/HoneyTipPostService.java
@@ -3,8 +3,8 @@ package com.api.ttoklip.domain.honeytip.post.service;
 import static com.api.ttoklip.global.util.SecurityUtil.getCurrentMember;
 
 import com.api.ttoklip.domain.common.Category;
-import com.api.ttoklip.domain.common.filtering.aop.annotation.CheckBadWordCreate;
-import com.api.ttoklip.domain.common.filtering.aop.annotation.CheckBadWordUpdate;
+import com.api.ttoklip.domain.aop.filtering.annotation.CheckBadWordCreate;
+import com.api.ttoklip.domain.aop.filtering.annotation.CheckBadWordUpdate;
 import com.api.ttoklip.domain.common.report.dto.ReportCreateRequest;
 import com.api.ttoklip.domain.common.report.service.ReportService;
 import com.api.ttoklip.domain.honeytip.comment.domain.HoneyTipComment;
