+++ src/main/java/com/api/ttoklip/domain/town/cart/post/service/CartPostService.java
@@ -2,13 +2,13 @@ package com.api.ttoklip.domain.town.cart.post.service;
 
 import static com.api.ttoklip.global.util.SecurityUtil.getCurrentMember;
 
-import com.api.ttoklip.domain.common.filtering.aop.annotation.CheckBadWordCreate;
-import com.api.ttoklip.domain.common.filtering.aop.annotation.CheckBadWordUpdate;
+import com.api.ttoklip.domain.aop.filtering.annotation.CheckBadWordCreate;
+import com.api.ttoklip.domain.aop.filtering.annotation.CheckBadWordUpdate;
 import com.api.ttoklip.domain.common.report.dto.ReportCreateRequest;
 import com.api.ttoklip.domain.common.report.service.ReportService;
 import com.api.ttoklip.domain.member.domain.Member;
 import com.api.ttoklip.domain.mypage.dto.response.UserCartSingleResponse;
-import com.api.ttoklip.domain.notification.aop.annotation.SendNotification;
+import com.api.ttoklip.domain.aop.notification.annotation.SendNotification;
 import com.api.ttoklip.domain.town.cart.comment.CartComment;
 import com.api.ttoklip.domain.town.cart.image.service.CartImageService;
 import com.api.ttoklip.domain.town.cart.itemUrl.service.ItemUrlService;
